👌 Further clean up main function
